+++ subgrade-sdk/src/main/java/com/dwarfeng/subgrade/sdk/interceptor/permission/PermissionRequiredAdvisor.java
@@ -1,5 +1,6 @@
 package com.dwarfeng.subgrade.sdk.interceptor.permission;
 
+import com.dwarfeng.subgrade.sdk.SystemPropertyConstants;
 import com.dwarfeng.subgrade.sdk.interceptor.AdvisorUtil;
 import com.dwarfeng.subgrade.stack.handler.PermissionHandler;
 import org.aspectj.lang.ProceedingJoinPoint;
@@ -8,17 +9,55 @@ import org.aspectj.lang.annotation.Aspect;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.core.Ordered;
 import org.springframework.core.annotation.Order;
+import org.springframework.expression.ExpressionParser;
+import org.springframework.expression.ParserContext;
 import org.springframework.stereotype.Component;
 
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-import java.util.Objects;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import java.util.*;
 
 /**
  * 权限需求增强。
+ *
+ * <p>
+ * 该类是一个 AOP 增强类，用于处理方法上的 {@link PermissionRequired} 注解。
+ * 该类将 {@link PermissionRequired} 注解中的 {@link PermissionRequired#value()}
+ * 中定义的权限与当前用户的权限进行比较，如果用户缺失所需的权限，则调用 {@link PermissionRequiredAopManager} 的相关调度方法。
+ *
+ * <p>
+ * 该类处理 {@link PermissionRequired} 注解时，基于部分系统属性的配置，
+ * 可对注解中的 {@link PermissionRequired#value()} 应用不同的处理逻辑。<br>
+ * 系统属性基于优先级进行判断，该类将按照优先级从高到低的顺序判断系统属性，如果高优先级的系统属性处于启用状态，
+ * 则使用该系统属性对应的处理逻辑；如果高优先级的系统属性处于禁用状态，则继续判断下一个优先级的系统属性，
+ * 直到找到一个启用状态的系统属性或者所有系统属性都处于禁用状态。<br>
+ * 相关的系统属性以及优先级见下表：
+ * <table>
+ *     <tr>
+ *         <th>优先级</th>
+ *         <th>系统属性</th>
+ *         <th>处理逻辑</th>
+ *     </tr>
+ *     <tr>
+ *         <td>1</td>
+ *         <td>{@link SystemPropertyConstants#VALUE_PERMISSION_REQUIRED_USE_SPEL}</td>
+ *         <td>
+ *             使用 SPEL 对 {@link PermissionRequired#value()} 中的每个权限值进行 SPEL 解析，
+ *             解析后的结果作为最终的权限值列表。
+ *         </td>
+ *     </tr>
+ *     <tr>
+ *         <td>2</td>
+ *         <td>无（缺省）</td>
+ *         <td>
+ *             直接将 {@link PermissionRequired#value()} 中的权限值作为最终的权限值解析，
+ *             解析后的结果作为最终的权限值列表。
+ *         </td>
+ *     </tr>
+ * </table>
  *
  * @author DwArFeng
  * @since 1.7.0
@@ -30,7 +69,19 @@ import java.util.Objects;
 @Order(Ordered.HIGHEST_PRECEDENCE + 1100)
 public class PermissionRequiredAdvisor {
 
-    private final Logger LOGGER = LoggerFactory.getLogger(PermissionRequiredAdvisor.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(PermissionRequiredAdvisor.class);
+
+    /**
+     * 是否启用权限需求注解的 SPEL 表达式解析功能。
+     *
+     * @since 1.7.3
+     */
+    private static final boolean PERMISSION_REQUIRED_USE_SPEL = Boolean.parseBoolean(
+            System.getProperty(
+                    SystemPropertyConstants.VALUE_PERMISSION_REQUIRED_USE_SPEL,
+                    Boolean.FALSE.toString()
+            )
+    );
 
     // 为了与旧代码兼容，使用字段注入，忽略相关警告。
     @SuppressWarnings("SpringJavaAutowiredFieldsWarningInspection")
@@ -42,6 +93,42 @@ public class PermissionRequiredAdvisor {
     @Autowired
     private PermissionRequiredAopManager manager;
 
+    /**
+     * 可配置 bean 工厂。
+     *
+     * <p>
+     * 仅当解析方式为 SPEL 时，需要该属性。
+     *
+     * @since 1.7.3
+     */
+    // 为了与旧代码兼容，使用字段注入，忽略相关警告。
+    @SuppressWarnings("SpringJavaAutowiredFieldsWarningInspection")
+    @Nullable
+    @Autowired(required = false)
+    private ConfigurableBeanFactory configurableBeanFactory;
+
+    /**
+     * SPEL 表达式解析器。
+     *
+     * @since 1.7.3
+     */
+    // 为了与旧代码兼容，使用字段注入，忽略相关警告。
+    @SuppressWarnings("SpringJavaAutowiredFieldsWarningInspection")
+    @Nullable
+    @Autowired(required = false)
+    private ExpressionParser expressionParser;
+
+    /**
+     * SPEL 解析上下文。
+     *
+     * @since 1.7.3
+     */
+    // 为了与旧代码兼容，使用字段注入，忽略相关警告。
+    @SuppressWarnings("SpringJavaAutowiredFieldsWarningInspection")
+    @Nullable
+    @Autowired(required = false)
+    private ParserContext parserContext;
+
     @Around("@annotation(com.dwarfeng.subgrade.sdk.interceptor.permission.PermissionRequired)")
     public Object around(ProceedingJoinPoint pjp) throws Throwable {
         // 记录日志。
@@ -51,39 +138,106 @@ public class PermissionRequiredAdvisor {
         String userId = manager.getUserId(pjp);
         LOGGER.debug("manager.getUserId(pjp) = {}", userId);
 
-        // 获取方法所需的执行权限。
+        //获取方法所需的执行权限。
         LOGGER.debug("扫描 @PermissionRequired 注解，获取方法执行所需的权限");
         PermissionRequired permissionRequired = AdvisorUtil.directMethodAnnotation(pjp, PermissionRequired.class);
-        LOGGER.debug("所需的权限: ");
         for (String s : permissionRequired.value()) {
             LOGGER.debug("\t{}", s);
         }
-        List<String> permissionList = Arrays.asList(permissionRequired.value());
-
-        /*
-         * 在一般情况下，hasPermission 的实现效率要远高于 getMissingPermissions 的实现效率，
-         * 因此先调用 hasPermission 方法，如果 hasPermission 方法返回 false，则调用 getMissingPermissions 方法。
-         */
-        boolean hasAllPermission = permissionHandler.hasPermission(userId, permissionList);
-        LOGGER.debug("permissionHandler.hasPermission(userId, permissionList) = {}", hasAllPermission);
-        if (hasAllPermission) {
-            LOGGER.debug("用户具有所需的权限，直接调度原始方法...");
-            return pjp.proceed(pjp.getArgs());
+        // 解析权限值列表（可能包含 SPEL 表达式）。
+        List<String> permissionList = resolvePermissions(permissionRequired.value());
+        LOGGER.debug("所需的权限: ");
+        for (int i = 0; i < permissionRequired.value().length; i++) {
+            LOGGER.debug("\t{} -> {}", permissionRequired.value()[i], permissionList.get(i));
         }
 
-        // 代码执行到此处意味着用户缺失权限，调用 getMissingPermissions 方法获取用户缺失的权限。
+        // 获取用户缺失的权限。
         LOGGER.debug("查询用户缺失的权限...");
         List<String> missingPermissions = permissionHandler.getMissingPermissions(userId, permissionList);
-        if (Objects.isNull(missingPermissions)) {
-            missingPermissions = Collections.emptyList();
+
+        // 如果用户缺失权限，则调用 manager 的相关调度方法。
+        if (!missingPermissions.isEmpty()) {
+            LOGGER.debug("用户缺失权限，调用 manager 的相关调度方法...");
+            for (String s : missingPermissions) {
+                LOGGER.debug("\t{}", s);
+            }
+            return manager.onMissingPermission(pjp, userId, missingPermissions);
         }
 
-        // 调用 manager 的相关调度方法。
-        LOGGER.debug("用户缺失权限，调用 manager 的相关调度方法...");
-        for (String s : missingPermissions) {
-            LOGGER.debug("\t{}", s);
+        // 原始方法的调度。
+        return pjp.proceed(pjp.getArgs());
+    }
+
+    /**
+     * 解析权限值列表。
+     *
+     * <p>
+     * 根据系统属性 {@link SystemPropertyConstants#VALUE_PERMISSION_REQUIRED_USE_SPEL} 决定是否启用 SPEL 表达式解析。
+     * 如果启用 SPEL 解析，则对每个权限值进行 SPEL 表达式解析；否则直接返回原始权限值列表。
+     *
+     * @param permissionValues 原始权限值数组。
+     * @return 解析后的权限值列表。
+     * @since 1.7.3
+     */
+    private List<String> resolvePermissions(@Nonnull String[] permissionValues) {
+        if (PERMISSION_REQUIRED_USE_SPEL) {
+            LOGGER.debug("SPEL 解析功能已启用, 开始进行 SPEL 解析");
+            return resolvePermissionsAsSpel(permissionValues);
+        } else {
+            LOGGER.debug("SPEL 解析功能已禁用, 直接返回原始权限值列表");
+            return resolvePermissionsAsIs(permissionValues);
         }
-        return manager.onMissingPermission(pjp, userId, missingPermissions);
+    }
+
+    private List<String> resolvePermissionsAsIs(String[] permissionValues) {
+        return Arrays.asList(permissionValues);
+    }
+
+    private List<String> resolvePermissionsAsSpel(String[] permissionValues) {
+        // 确保可配置 bean 工厂不为 null。
+        if (Objects.isNull(configurableBeanFactory)) {
+            LOGGER.warn("可配置 bean 工厂未设置, 无法进行 SPEL 解析, 将抛出异常");
+            LOGGER.warn(
+                    "如果希望禁用 SPEL 解析功能, 请将系统属性 {} 设为 false",
+                    SystemPropertyConstants.VALUE_PERMISSION_REQUIRED_USE_SPEL
+            );
+            throw new IllegalStateException("可配置 bean 工厂未设置, 无法进行 SPEL 解析");
+        }
+        // 确保表达式解析器不为 null。
+        if (Objects.isNull(expressionParser)) {
+            LOGGER.warn("表达式解析器未设置, 无法进行 SPEL 解析, 将抛出异常");
+            LOGGER.warn(
+                    "如果希望禁用 SPEL 解析功能, 请将系统属性 {} 设为 false",
+                    SystemPropertyConstants.VALUE_PERMISSION_REQUIRED_USE_SPEL
+            );
+            throw new IllegalStateException("表达式解析器未设置, 无法进行 SPEL 解析");
+        }
+        // 确保解析上下文不为 null。
+        if (Objects.isNull(parserContext)) {
+            LOGGER.warn("解析上下文未设置, 无法进行 SPEL 解析, 将抛出异常");
+            LOGGER.warn(
+                    "如果希望禁用 SPEL 解析功能, 请将系统属性 {} 设为 false",
+                    SystemPropertyConstants.VALUE_PERMISSION_REQUIRED_USE_SPEL
+            );
+            throw new IllegalStateException("解析上下文未设置, 无法进行 SPEL 解析");
+        }
+
+        // 创建解析后的权限值列表。
+        List<String> resolvedPermissions = new ArrayList<>();
+
+        // 对每个 permissionValues 中的值进行 SPEL 解析。
+        for (String permissionValue : permissionValues) {
+            // 解析占位符。
+            permissionValue = Optional.ofNullable(configurableBeanFactory.resolveEmbeddedValue(permissionValue))
+                    .orElse(permissionValue);
+            // 解析 SPEL。
+            permissionValue = expressionParser.parseExpression(permissionValue, parserContext).getValue(String.class);
+            // 将解析结果添加到结果列表中。
+            resolvedPermissions.add(permissionValue);
+        }
+
+        // 返回结果。
+        return resolvedPermissions;
     }
 
     public PermissionHandler getPermissionHandler() {
@@ -100,5 +254,43 @@ public class PermissionRequiredAdvisor {
 
     public void setManager(PermissionRequiredAopManager manager) {
         this.manager = manager;
+    }
+
+    @Nullable
+    public ConfigurableBeanFactory getConfigurableBeanFactory() {
+        return configurableBeanFactory;
+    }
+
+    public void setConfigurableBeanFactory(@Nullable ConfigurableBeanFactory configurableBeanFactory) {
+        this.configurableBeanFactory = configurableBeanFactory;
+    }
+
+    @Nullable
+    public ExpressionParser getExpressionParser() {
+        return expressionParser;
+    }
+
+    public void setExpressionParser(@Nullable ExpressionParser expressionParser) {
+        this.expressionParser = expressionParser;
+    }
+
+    @Nullable
+    public ParserContext getParserContext() {
+        return parserContext;
+    }
+
+    public void setParserContext(@Nullable ParserContext parserContext) {
+        this.parserContext = parserContext;
+    }
+
+    @Override
+    public String toString() {
+        return "PermissionRequiredAdvisor{" +
+                "permissionHandler=" + permissionHandler +
+                ", manager=" + manager +
+                ", configurableBeanFactory=" + configurableBeanFactory +
+                ", expressionParser=" + expressionParser +
+                ", parserContext=" + parserContext +
+                '}';
     }
 }